🚚 move: MainGlobalNavBar, FootNavBar 파일 이동
